requirements.txt
fastapi
uvicorn[standard]
httpx[socks]
psycopg2-binary
python-dotenv

